requirements.txt
pandas==2.2.2
numpy==2.0.1
scikit-learn==1.5.1
joblib==1.4.2
PyYAML==6.0.2
streamlit==1.36.0
matplotlib==3.9.0
shap==0.45.1
uvicorn==0.30.1
fastapi==0.111.0
xgboost==3.0.4
kagglehub==0.3.13
